refactor: extract request form into component
